+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Option to configure the timeout and keep alive interval for the MQTT
+  connection to astarte.
+
 ## [0.5.2] - 2023-07-03
 ### Added
 -  Add support to receive `device_id` option from dbus.